[level 2] Title: n^2 배열 자르기, Time: 7.46 ms, Memory: 108 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -4,7 +4,7 @@ class Solution {
         int[] answer = new int[difference + 1];
         int idx = 0;
         for (long i = left; i <= right; i++) {
-            int row = (int) (i / (long) n);
+            int row = (int) (i / n);
             int col = (int) (i % (long) n);
             
             if (col <= row) {
